Address review feedback: fold CLI driver back into main.cpp, merge CI, drop Jupyter, avoid List allocs in File line reads
