+++ codon/parser/common.cpp
@@ -66,7 +66,6 @@ ImportFile IFilesystem::get_root(const path_t &sp) const {
              "bad path substitution: {}, {}", s, root);
   auto module = s.substr(root.size() + 1, s.size() - root.size() - ext.size() - 1);
   std::ranges::replace(module, '/', '.');
-  std::ranges::replace(module, '\\', '.'); // Windows path separator -> module dot
   return ImportFile{(!isStdLib && root == module0) ? ImportFile::PACKAGE
                                                    : ImportFile::STDLIB,
                     s, module};
